Moved all code to single file
